utils class and save file functionality
